# Optimize `Node::is_ancestor_of` for nodes in tree